Revert "bunch of stuff"
This reverts commit 51eb64f133097b436be3f2531950e60e482fc988.

+++ src/main/java/frc/robot/subsystems/FalconClosedLoop.java
@@ -33,7 +33,7 @@ public class FalconClosedLoop extends SubsystemBase {
     private ControlMode CurrentControlMode;
     private static PIDCoefficients VelocityPIDCoefficients = new PIDCoefficients(.23,0.0004,7,0);
     private static PIDCoefficients PositionPIDCoefficients = new PIDCoefficients(1,0,0,0);
-    private double inverted = 1.0;//positive or negative
+    private int inverted = 1;//positive or negative
 
     public FalconClosedLoop(int talonId,int PIDLoopId,int timeoutMs,ControlMode ClosedLoopMode) {
         this.PIDLoopId = PIDLoopId;
@@ -95,9 +95,9 @@ public class FalconClosedLoop extends SubsystemBase {
     }
     public void setInverted(Boolean invert){
         if (invert){
-            this.inverted = -1.0;
+            this.inverted = -1;
         } else if (!invert){
-            this.inverted = 1.0;
+            this.inverted = 1;
         }else {
             throw new IllegalArgumentException("IDK why but this was supposed to be boolean in Falcon Closed loop set inverted")
         }
